canvas: removed dotted_* functions

+++ tcob/include/tcob/core/Color.inl
@@ -119,10 +119,10 @@ auto constexpr color::as_grayscale(f32 redFactor, f32 greenFactor, f32 blueFacto
 
 auto constexpr color::Lerp(color left, color right, f64 step) -> color
 {
-    u8 const nr {static_cast<u8>(left.R + static_cast<f64>((right.R - left.R) * step))};
-    u8 const ng {static_cast<u8>(left.G + static_cast<f64>((right.G - left.G) * step))};
-    u8 const nb {static_cast<u8>(left.B + static_cast<f64>((right.B - left.B) * step))};
-    u8 const na {static_cast<u8>(left.A + static_cast<f64>((right.A - left.A) * step))};
+    u8 const nr {static_cast<u8>(left.R + (right.R - left.R) * step)};
+    u8 const ng {static_cast<u8>(left.G + (right.G - left.G) * step)};
+    u8 const nb {static_cast<u8>(left.B + (right.B - left.B) * step)};
+    u8 const na {static_cast<u8>(left.A + (right.A - left.A) * step)};
     return {nr, ng, nb, na};
 }
